Moved RemoveParameters from the RefactorMenu into a Command class, see #762
