--- RetailCoder.VBE/UI/Command/RefactorRemoveParametersCommand.cs
@@ -1,15 +1,0 @@
-using System;
-using System.Runtime.InteropServices;
-using System.Windows.Input;
-
-namespace Rubberduck.UI.Command
-{
-    [ComVisible(false)]
-    public class RefactorRemoveParametersCommand : CommandBase
-    {
-        public override void Execute(object parameter)
-        {
-            throw new NotImplementedException();
-        }
-    }
-}
