+++ RetailCoder.VBE/UI/Command/Refactorings/RefactorRemoveParametersCommand.cs
@@ -1,0 +1,43 @@
+using Microsoft.Vbe.Interop;
+using Rubberduck.Parsing;
+using Rubberduck.VBEditor;
+using Rubberduck.VBEditor.VBEInterfaces.RubberduckCodePane;
+using System.Runtime.InteropServices;
+using Rubberduck.Refactorings.RemoveParameters;
+using Rubberduck.UI.Refactorings;
+
+namespace Rubberduck.UI.Command.Refactorings
+{
+    [ComVisible(false)]
+    public class RefactorRemoveParametersCommand : RefactorCommandBase
+    {
+        private readonly ICodePaneWrapperFactory _wrapperWrapperFactory;
+
+        public RefactorRemoveParametersCommand(VBE ide, IRubberduckParser parser, IActiveCodePaneEditor editor, ICodePaneWrapperFactory wrapperWrapperFactory) 
+            : base (ide, parser, editor)
+        {
+            _wrapperWrapperFactory = wrapperWrapperFactory;
+        }
+
+        public override void Execute(object parameter)
+        {
+            if (_ide.ActiveCodePane == null)
+            {
+                return;
+            }
+            var codePane = _wrapperWrapperFactory.Create(_ide.ActiveCodePane);
+            var selection = new QualifiedSelection(new QualifiedModuleName(codePane.CodeModule.Parent), codePane.Selection);
+            // duplicates ReorderParameters Implementation until here... extract common method?
+            // TryGetQualifiedSelection?
+            var progress = new ParsingProgressPresenter();
+            var result = progress.Parse(_parser, _ide.ActiveVBProject);
+
+            using (var view = new RemoveParametersDialog())
+            {
+                var factory = new RemoveParametersPresenterFactory(_editor, view, result, new MessageBox());
+                var refactoring = new RemoveParametersRefactoring(factory, _editor);
+                refactoring.Refactor(selection);
+            }
+        }
+    }
+}
